Moved FWCore/SOA tests to catch2
