+++ FWCore/SOA/test/table_filling_t.catch2.cpp
@@ -2,28 +2,7 @@
 #include "FWCore/SOA/interface/TableView.h"
 #include "FWCore/SOA/interface/Column.h"
 
-#include <cppunit/extensions/HelperMacros.h>
-
-class testTableFilling : public CppUnit::TestFixture {
-  CPPUNIT_TEST_SUITE(testTableFilling);
-
-  CPPUNIT_TEST(soaDeclareDefaultTest1);
-  CPPUNIT_TEST(soaDeclareDefaultTest2);
-  CPPUNIT_TEST(soaDeclareDefaultTest3);
-  CPPUNIT_TEST(soaDeclareDefaultTest4);
-  CPPUNIT_TEST(soaDeclareDefaultTest5);
-  CPPUNIT_TEST_SUITE_END();
-
-public:
-  void setUp() {}
-  void tearDown() {}
-
-  void soaDeclareDefaultTest1();
-  void soaDeclareDefaultTest2();
-  void soaDeclareDefaultTest3();
-  void soaDeclareDefaultTest4();
-  void soaDeclareDefaultTest5();
-};
+#include "catch2/catch_all.hpp"
 
 namespace ts {
 
@@ -57,16 +36,13 @@ namespace edm::soa {
 
 }  // namespace edm::soa
 
-///registration of the test so that the runner can find it
-CPPUNIT_TEST_SUITE_REGISTRATION(testTableFilling);
-
 namespace {
   template <class Object>
   void validateEtaPhiTable(ts::EtaPhiTableView table, std::vector<Object> const& objects) {
     int iRow = 0;
     for (auto const& row : table) {
-      CPPUNIT_ASSERT(row.get<ts::col::Eta>() == objects[iRow].eta_);
-      CPPUNIT_ASSERT(row.get<ts::col::Phi>() == objects[iRow].phi_);
+      REQUIRE(row.get<ts::col::Eta>() == objects[iRow].eta_);
+      REQUIRE(row.get<ts::col::Phi>() == objects[iRow].phi_);
       ++iRow;
     }
   }
@@ -82,15 +58,6 @@ struct JetType1 {
   const float phi_;
 };
 
-void testTableFilling::soaDeclareDefaultTest1() {
-  std::vector<JetType1> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
-  std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
-
-  ts::EtaPhiTable table(jets);
-
-  validateEtaPhiTable(table, jets);
-}
-
 struct JetType2 {
   // jet type that is NOT compatible with the SOA_DECLARE_DEFAULT-generated
   // value_for_column functions...
@@ -103,15 +70,6 @@ struct JetType2 {
 // ...so we need to write our own value_for_column function for JetType2
 double value_for_column(JetType2 const& iJ, ts::col::Phi*) { return iJ.phi_; }
 
-void testTableFilling::soaDeclareDefaultTest2() {
-  std::vector<JetType2> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
-  std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
-
-  ts::EtaPhiTable table(jets);
-
-  validateEtaPhiTable(table, jets);
-}
-
 namespace ts::reco {
 
   struct JetType3 {
@@ -123,17 +81,6 @@ namespace ts::reco {
 
   double value_for_column(ts::reco::JetType3 const& iJ, ts::col::Phi*) { return iJ.phi_; }
 }  // namespace ts::reco
-
-// same tests as for JetType2 but with the type and the value_for_column
-// defined in their own test namespace
-void testTableFilling::soaDeclareDefaultTest3() {
-  std::vector<ts::reco::JetType3> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
-  std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
-
-  ts::EtaPhiTable table(jets);
-
-  validateEtaPhiTable(table, jets);
-}
 
 struct JetType4 {
   // jet type that is compatible with the SOA_DECLARE_DEFAULT-generated
@@ -148,15 +95,6 @@ struct JetType4 {
 };
 
 double value_for_column(JetType4 const& iJ, ts::col::Phi*) { return iJ.phi_; }
-
-void testTableFilling::soaDeclareDefaultTest4() {
-  std::vector<JetType4> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
-  std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
-
-  ts::EtaPhiTable table(jets);
-
-  validateEtaPhiTable(table, jets);
-}
 
 namespace ts::reco {
 
@@ -176,13 +114,49 @@ namespace ts::reco {
 
 }  // namespace ts::reco
 
-void testTableFilling::soaDeclareDefaultTest5() {
-  std::vector<ts::reco::JetType5> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
-  std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
+TEST_CASE("TableFilling", "[SOA]") {
+  SECTION("soaDeclareDefaultTest1") {
+    std::vector<JetType1> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
+    std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
 
-  ts::EtaPhiTable table(jets);
+    ts::EtaPhiTable table(jets);
 
-  validateEtaPhiTable(table, jets);
+    validateEtaPhiTable(table, jets);
+  }
+
+  SECTION("soaDeclareDefaultTest2") {
+    std::vector<JetType2> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
+    std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
+
+    ts::EtaPhiTable table(jets);
+
+    validateEtaPhiTable(table, jets);
+  }
+
+  SECTION("soaDeclareDefaultTest3") {
+    std::vector<ts::reco::JetType3> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
+    std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
+
+    ts::EtaPhiTable table(jets);
+
+    validateEtaPhiTable(table, jets);
+  }
+
+  SECTION("soaDeclareDefaultTest4") {
+    std::vector<JetType4> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
+    std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
+
+    ts::EtaPhiTable table(jets);
+
+    validateEtaPhiTable(table, jets);
+  }
+
+  SECTION("soaDeclareDefaultTest5") {
+    std::vector<ts::reco::JetType5> jets = {{1., 3.14}, {2., 0.}, {4., 1.3}};
+    std::vector<std::string> labels = {{"jet0", "jet1", "jet2"}};
+
+    ts::EtaPhiTable table(jets);
+
+    validateEtaPhiTable(table, jets);
+  }
 }
-
-#include <Utilities/Testing/interface/CppUnit_testdriver.icpp>
